Fix DataUtils memory leak
